rename tag specs
