--- components/Tag/react/Tag.cy.tsx
@@ -1,13 +1,0 @@
-/// <reference types="cypress" />
-
-import * as React from 'react'
-import { mount } from 'cypress/react18'
-import TagStory from './Tag.rootstory'
-import assertions from '../assertions'
-
-describe('Tag', () => {
-  function mountStory() {
-    mount(<TagStory />)
-  }
-  assertions(mountStory)
-})
