+++ components/Tag/react/TagReact.cy.tsx
@@ -1,18 +1,15 @@
-import clsx from 'clsx'
-import Tag from './Tag'
-import * as React from 'react'
-import { SizeClasses, ColorClasses } from '../constants'
+/// <reference types="cypress" />
 
-export default () => (
+import * as React from 'react'
+import { mount } from 'cypress/react18'
+import assertions from '../assertions'
+
+const TagStory = () => (
   <div className="flex flex-row flex-wrap items-start justify-center bg-gray-700 gap-6">
     {(Object.keys(SizeClasses) as Array<keyof typeof SizeClasses>).map(
       (size) => {
         return (
-          <div
-            className={clsx(
-              'flex flex-col items-center gap-3 justify-center my-4 p-4 bg-white rounded-lg'
-            )}
-          >
+          <div className="flex flex-col items-center gap-3 justify-center my-4 p-4 bg-white rounded-lg">
             <h3 className="text-right">{size}</h3>
             {(
               Object.keys(ColorClasses) as Array<keyof typeof ColorClasses>
@@ -38,3 +35,10 @@ export default () => (
     )}
   </div>
 )
+
+describe('Tag', () => {
+  function mountStory() {
+    mount(<TagStory />)
+  }
+  assertions(mountStory)
+})
